feat(api): add verifyDeveloper middleware
