+++ apps/api/src/interfaces/middleware.interface.ts
@@ -1,9 +1,12 @@
 import { Request } from 'express';
 
+export type RoleType = 'USER' | 'ADMIN' | 'DEVELOPER';
+
 export interface UserToken {
   id: string;
   email: string;
   jobId: string;
+  role: RoleType;
 }
 
 export interface UserRequest extends Request {
